requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.6
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
PyJWT==2.6.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tzdata==2022.7
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
